refactor: move credentials to separate module
- Create mixpanel/credentials.py for authentication classes
- Move ServiceAccountCredentials and MixpanelCredentials to new file
- Import from credentials module in __init__.py
- Improves code organization and separation of concerns

Benefits:
- Cleaner module structure
- Easier to add new credential types in one place
- Credentials logic isolated from core tracking code
